Rename the shared theme module to theme-precedence
`user-intent-theme` read as an awkward noun phrase. The module's subject is the
precedence between styling set on a figure and styling supplied by a theme, so
name it for that.

Kept a descriptive name rather than `theme-utils`. Streamlit's own frontend does
use utils modules, including `theme/utils.ts`, but the pattern there is that
specific single-purpose things get descriptive names -- `getColors`,
`getShadows`, `namedColors` -- and `utils` holds the residue. This is the former.
Kebab-case follows this repo rather than streamlit's camelCase.

The export stays `withUserIntent`, which reads better at the call site than the
file's concern would: `use_theme(withUserIntent(theme))`.
