+++ streamlit_bokeh/frontend/src/v1/index.ts
@@ -15,7 +15,7 @@
  */
 
 import { Streamlit, RenderData, Theme } from "streamlit-component-lib"
-import { withUserIntent } from "../shared/user-intent-theme"
+import { withUserIntent } from "../shared/theme-precedence"
 import { streamlitTheme } from "./streamlit-theme"
 
 declare global {
@@ -96,7 +96,7 @@ export const setChartThemeGenerator = () => {
         use_theme(null)
       } else if (resolvedTheme === "streamlit") {
         // Wrapped so the theme cannot suppress a visual whose colour the user
-        // set explicitly. See shared/user-intent-theme.ts.
+        // set explicitly. See shared/theme-precedence.ts.
         use_theme(withUserIntent(streamlitTheme(newAppTheme)))
       } else {
         use_theme(withUserIntent(builtInThemes[resolvedTheme]))
